Avoid running test suite with doc only changes
Note: Sometimes when PRs are rebased and force pushed the base of the
commit range may reference a commit which existed before the rebase.
Typically this isn't a problem but since CI repos start from scratch
that commit will probably not exist. When that occurs we can not
determine what files were changed so we automatically fall back to
running the test suite.